agent/requirements.txt
osquery
grpcio==1.51.1
grpcio-tools==1.51.1
configparser
PyJWT